requirements_new.txt
cython==0.29.34
keyboard==0.13.5
matplotlib==3.5.2
networkx==3.0
numpy==1.22.4
pandas==1.3.5
PyOpenGL==3.1.6
PyQt5==5.15.6
pyqtgraph==0.12.4
scikit_learn==1.1.1
scipy==1.8.1
seaborn==0.12.2
shapely==2.0.1
torch==1.13.1
tqdm==4.65.0
trimesh==3.12.5
wakepy==0.6.0

